colSumamrize {preprocessCore}	R Documentation

Summarize the column of matrices

Description

Compute column wise summary values of a matrix.

Usage

colSummarizeAvg(y)
colSummarizeAvgLog(y)
colSummarizeBiweight(y)
colSummarizeBiweightLog(y)
colSummarizeLogAvg(y)
colSummarizeLogMedian(y)
colSummarizeMedian(y)
colSummarizeMedianLog(y)
colSummarizeMedianpolish(y)
colSummarizeMedianpolishLog(y)

Arguments

y
A numeric matrix

Details

This groups of functions summarize the columns of a given matrices.

colSummarizeAvgTake means in column-wise manner

colSummarizeAvgLoglog2 transform the data and then take means in column-wise manner

colSummarizeBiweightSummarize each column using a one step Tukey Biweight procedure

colSummarizeBiweightLoglog2 transform the data and then summarize each column using a one step Tuke Biweight procedure

colSummarizeLogAvgCompute the mean of each column and then log2 transform it

colSummarizeLogMedianCompute the median of each column and then log2 transform it

colSummarizeMedianCompute the median of each column

colSummarizeMedianLoglog2 transform the data and then summarize each column using the median

colSummarizeMedianpolishUse the median polish to summarize each column, by also using a row effect (not returned)

colSummarizeMedianpolishLoglog2 transform the data and then use the median polish to summarize each column, by also using a row effect (not returned)

Value

A list with following items:

Estimates
Summary values for each column.

StdErrors
Standard error estimates.

Author(s)

B. M. Bolstad bmb@bmbolstad.com

Examples

y <- matrix(10+rnorm(100),20,5)

colSummarizeAvg(y)
colSummarizeAvgLog(y)
colSummarizeBiweight(y)
colSummarizeBiweightLog(y)
colSummarizeLogAvg(y)
colSummarizeLogMedian(y)
colSummarizeMedian(y)
colSummarizeMedianLog(y)
colSummarizeMedianpolish(y)
colSummarizeMedianpolishLog(y)

[Package preprocessCore version 1.52.1 Index]