rma.background.correct {preprocessCore}	R Documentation

RMA Background Correction

Description

Background correct each column of a matrix

Usage

  rma.background.correct(x,copy=TRUE)

Arguments

x
A matrix of intensities where each column corresponds to a chip and each row is a probe.

copy
Make a copy of matrix before background correctiong. Usually safer to work with a copy, but in certain situations not making a copy of the matrix, but instead background correcting it in place will be more memory friendly.

Details

Assumes PMs are a convolution of normal and exponentional. So we observe X+Y where X is backround and Y is signal. bg.adjust returns E[Y|X+Y, Y>0] as our backround corrected PM.

Value

A RMA background corrected matrix.

Author(s)

Ben Bolstad, bmbolstad.com

References

Bolstad, BM (2004) Low Level Analysis of High-density Oligonucleotide Array Data: Background, Normalization and Summarization. PhD Dissertation. University of California, Berkeley. pp 17-21

[Package preprocessCore version 1.52.1 Index]